code show list internship class, code delete internship class

+++ app/Http/Controllers/InternshipclassController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Group;
 use App\Internshipclass;
 use Illuminate\Http\Request;
 
@@ -14,8 +15,8 @@ class InternshipclassController extends Controller
      */
     public function index()
     {
-        $danhsachdot = Internshipclass::all();
-        return view('admin.pages.internshipClass.index', ['danhsachdot'=>$danhsachdot]);
+        $listClass = Internshipclass::all();
+        return view('admin.pages.internshipClass.index', ['listClass'=>$listClass]);
     }
 
     /**
@@ -82,6 +83,11 @@ class InternshipclassController extends Controller
     public function destroy($id)
     {
         $a = Internshipclass::find($id);
+        $group = Group::where('class_id', $a->id)->get();
+        $count = count($group);
+        if($count != 0){
+            return redirect('admin/internshipClass')->with('fail', 'Xóa không thành công.');
+        }
         $a->delete();
         return redirect('admin/internshipClass')->with('success', 'Xóa thành công');
     }
